:star: Modularize app.js & converting some jQuery to vanilla JS
