+++ src/js/copyToUsername.js.vaniquerycache
@@ -1,0 +1,37 @@
+/**
+ * COPY-TO-USERNAME CHECKBOX FUNCTIONALITIES
+ */
+
+'use strict';
+
+const copyToUsername = () => {
+  $('#copy-to-username').click(
+    /** @this HTMLElement */
+    function () {
+      $('#target-username').val(() =>
+        $(this).prop('checked') ? $('#target-owner').val() : ''
+      );
+    }
+  );
+}
+
+/**
+ * Copy target-owner to target-username if 'copy-to-username' is checked
+ */
+const copyToTargetOwner = () => {
+  $('#target-owner').keyup(() => {
+    $('#copy-to-username').prop('checked') &&
+      $('#target-username').val($('#target-owner').val());
+  })
+};
+
+
+$('#target-username').keyup(
+  /** @this HTMLElement */
+  function () {
+    $('#copy-to-username').prop('checked', () => {
+      return $(this).val() === $('#target-owner').val();
+    });
+  }
+);
+
