+++ src/js/manipulateEntries.js.vaniquerycache
@@ -1,0 +1,62 @@
+/**
+ * 
+ */
+
+'use strict';
+
+const clickToDeleteAllEntries = (selector) => {
+  $(selector)
+    .children()
+    .each(
+      /** @this HTMLElement */
+      function () {
+        if (this.getAttribute('new') === 'true') {
+          this.remove();
+        } else {
+          this.children('.card').classList.add('deleted-card');
+          this.parentNode.querySelectorAll('.invalid-input').classList.add('hidden');
+          this.children('.recover-button').removeAttribute('disabled');
+          this.children('.delete-button').classList.add('hidden');
+          this.children('.recover-button').classList.remove('hidden');
+          this.setAttribute('data-todo', 'delete');
+        }
+      }
+    );
+};
+// Original code:
+// const clickToDeleteAllEntries = (selector) => {
+//   $(selector)
+//     .children()
+//     .each(
+//       /** @this HTMLElement */
+//       function () {
+//         if ($(this).attr('new') === 'true') {
+//           $(this).remove();
+//         } else {
+//           $(this).children('.card').addClass('deleted-card');
+//           $(this).parent().find('.invalid-input').addClass('hidden');
+//           $(this).children('.recover-button').removeAttr('disabled');
+//           $(this).children('.delete-button').addClass('hidden');
+//           $(this).children('.recover-button').removeClass('hidden');
+//           $(this).attr('data-todo', 'delete');
+//         }
+//       }
+//     );
+//   checkIfEnableCommit();
+// };
+
+const clickToDeleteAllLabels = () => {
+  $('#delete-all-labels').click(() => {
+    clickToDeleteAllEntries('#form-labels');
+    checkIfEnableCommit();
+  });
+};
+
+const clickToDeleteAllMilestones = () => {
+  $('#delete-all-milestones').click(() => {
+    clickToDeleteAllEntries('#form-milestones');
+    checkIfEnableCommit();
+  });
+};
+
+export { clickToDeleteAllLabels, clickToDeleteAllMilestones };
